+++ CHANGELOG.rst
@@ -39,6 +39,9 @@ Features
 *  **SendGrid:** Support both new "dynamic" and original "legacy" transactional
    templates. (See
    `docs <https://anymail.readthedocs.io/en/latest/esps/sendgrid/#sendgrid-templates>`__.)
+*  **SendGrid:** Allow merging `esp_extra["personalizations"]` dict into other message-derived
+   personalizations. (See
+   `docs <https://anymail.readthedocs.io/en/latest/esps/sendgrid/#sendgrid-esp-extra>`__.)
 
 
 v4.0